+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer_email "git@gilles.cornu.name"
 license          "Apache 2.0"
 description      "Installs sbt-extras to build scala projects with any sbt version you need"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.1.0"
+version          "0.2.0"
 
 depends          "java"
 conflicts        "typesafe-stack"