+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG for sbt-extras cookbook
 
+## 0.2.0:
+
+* Enhancement #3 - Add Templates for global config files (/etc/sbt/sbtopts and /etc/sbt/jvmopts)
+* Attributes modified (not backward compatible with 0.1.0)
+* Added timeout on 'execute' resources (sbt/scala downloads)
+
 ## 0.1.0:
 
 * Initial release